requirements.txt
requests
google-cloud-storage
google-cloud-secret-manager
google-ads
protobuf